Add files via upload
This is a draft. Please review and make changes. Please keep changes tracked, do not accept so I can keep track of all the changes. The Professor replied my email, and he wants us to add in-text and references. I will add references, and make some more changes to the document and provide a final draft by Friday evening.